Update the plugin to work with Android plugin 1.1.3
Note: This change is not backwards compatible with versions prior to
1.1.0. Some other language plugins jump through hoops to determine the
version being used, and then select the correct field dynamically. This
personally doesn't make sense to me, as the older version of the scala
plugin can continue to be used for those who need to use an older
version of the Gradle plugin.

I've also updated the readme as the `multiDexEnabled` option appears to
work just fine, now (and it is *ridiculously* fast after the first run,
as it's now incremental and doesn't need to calculate the main dex file)

Fixes #53